feat: use deno to parse git deps xml
I wanted to parse the xml file in but bazel doesn't support using
generated files in repo rules. To get around this I'm using deno to
parse the xml file and generate a json file that I can use to get the urls
for Unreal dependencies.

Deno works because it's a single binary that can be downloaded and it
automatically downloads js dependencies. I'm using the xml2js npm package
to parse the xml file.

I could probably use this directly to generate the urls I need.