wip: get the urls with the deno tool
This moves the deno parse tool to a new folder and adds a feature
to print the urls to stdout.

The next step is to download all the packs and calculate the hashes.
My rough plan is to use either golang or deno to download the packs, calculate
the sha256 hashes, and save it to a database in my cluster.