feat: combine db and deno to download dependencies
The deno now reads the sqlite db to get the sha256 of the
dependencies and creates a file for the repo rule to use to
download the dependencies. On first run, it ran in ~1300s and
on subsequent runs it ran in ~150s.

Deno has a sqlite module that can be used to read the db. I might
use it to write the db in the future. The go part is a bit
convuluted